move pose_vel operator to separate file
