+++ wave_gtsam/CMakeLists.txt
@@ -9,6 +9,7 @@ WAVE_ADD_MODULE(${PROJECT_NAME} DEPENDS
     src/hand_eye.cpp
     src/decaying_bias.cpp
     src/gps_factor_with_bias.cpp
+    src/pose_vel.cpp
     src/pose_vel_bias.cpp
     src/motion_factor.cpp
     src/localizer_state.cpp)
