+++ wave_gtsam/include/wave/gtsam/pose_vel.hpp
@@ -219,14 +219,4 @@ struct traits<wave::PoseVel> {
 };
 }
 
-namespace wave {
-
-PoseVel operator*(const PoseVel &m1, const PoseVel &m2) {
-    wave::PoseVel retval;
-    retval.pose = m1.pose * m2.pose;
-    retval.vel = m1.vel + m2.vel;
-    return retval;
-}
-}
-
 #endif  // WAVE_POSE_VEL_HPP
